fix: sonar issue Extract MultiSelectDropdown to a Separate Component
