+++ src/components/ModalDetails/ModalDetails.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useMemo, useCallback } from "react";
+import React, { useEffect, useState } from "react";
 import {
   Label,
   FormGroup,
@@ -15,7 +15,7 @@ import { assignAriaLabelsToDatePickerButtons } from "../../additional-functions/
 
 import { ArrowBackSharp } from "@material-ui/icons";
 import SelectBox from "../DetailsSelectBox/DetailsSelectBox";
-import MultiSelectCombobox from "../MultiSelectCombobox/MultiSelectCombobox";
+import MultiSelectDropdownWrapper from "../MultiSelectCombobox/MultiSelectCombooxWrapper.js";
 import {
   getMaxFutureDate,
   formatDateToInput
@@ -45,7 +45,6 @@ const ModalDetails = ({
     textOverflow: "ellipsis",
     overflow: "auto",
   };
-  const [comboBoxItems, setComboBoxItems] = useState([]);
   useEffect(() => {
     assignAriaLabelsToDatePickerButtons();
     let found = false;
@@ -235,71 +234,14 @@ const ModalDetails = ({
         break;
 
       case "multiSelectDropdown":
-        const [comboBoxItems, setComboBoxItems] = useState([]);
-
-        useEffect(() => {
-          if (value[6] && value[6].length > 0) {
-            const items = value[6]
-              .filter(valueItem => valueItem.code !== "")
-              .map(valueItem => ({
-                id: valueItem.code,
-                label: valueItem.name,
-                selected: !create && modalData?.[value[0]]
-                  ? modalData?.[value[0]].split(",").includes(valueItem.code)
-                  : false,
-                enabled: true,
-                disabled: valueItem.disabled || false
-              }));
-            setComboBoxItems(items);
-          }
-        }, [value[0], value[6], modalData, create]);
-
-        const selectedOptions = useMemo(() => {
-          return comboBoxItems
-            .filter(item => item.selected)
-            .map(item => item.id);
-        }, [comboBoxItems]);
-
-        const handleCodes = useCallback((id, updateType) => {
-          setComboBoxItems(prevItems => {
-            return prevItems.map(item => {
-              if (item.id === id) {
-                return { ...item, selected: updateType === "add" };
-              }
-              return item;
-            });
-          });
-        }, []);
-
-        useEffect(() => {
-          if (document.getElementById(value[1])) {
-            document.getElementById(value[1]).value = selectedOptions.join(",");
-          }
-        }, [selectedOptions, value[1]]);
-
-        let styles = {
-          listbox:
-            "list-box bg-white display-block height-15 width-card-lg overflow-y-scroll overflow-x-hidden border-top",
-          combobox:
-            "margin-top-1 margin-bottom-2 border-1px width-card-lg bg-white multi-select-combobox",
-        };
         comp = (
-          <div
-            className="modalUserInput"
-            value={selectedOptions.toString()}
-            id={value[1]}
-            epadataname={value[0]}
-          >
-            <MultiSelectCombobox
-              items={comboBoxItems}
-              entity={value[0]}
-              searchBy="contains"
-              onChangeUpdate={handleCodes}
-              styling={styles}
-              favicon={false}
-              disabled={isEditingDisabled}
-            />
-          </div>
+          <MultiSelectDropdownWrapper
+            value={value}
+            create={create}
+            modalData={modalData}
+            isEditingDisabled={isEditingDisabled}
+            cols={cols}
+          />
         );
         break;
